Create Discord-native scheduled events
Discord now supports scheduled events in the app itself, so we can
use those instead of relying on reactions to RSVP. This commit
also adds new config keys to set the server and voice channel that
games should be scheduled in. If the bot's to be expanded to
supporting multiple servers simultaneously, that'd need to be
stored in the DB, rather than static config, but it works fine for
now.